x & y compensation
When plane is tilted in xz and/or yz, respective x and/or y points are slightly offset. Otherwise (before this commit), the "normal plane" (aka image to engrave if 2d) would be scaled on x/y normal axes, depending on how much the normal plane is tilted on those axes.
This commit fixes that, moving x and y coordinates as well as z for a tilted plane.